Game 1
This game has a drag to match the color of the match rectangle. Drag the square to the color changer to change the color of the square. If the color changer is not the right color, move the square to the color changer changer to change the color of the color changer.

Co-Authored-By: Christopher Lendechy <53926684+ChristopherLendechy@users.noreply.github.com>